requirements.txt
python-dotenv==1.0.1
streamlit==1.32.2
pandas==2.2.1
sqlalchemy==2.0.27
requests==2.31.0
openpyxl==3.1.2
